move trpc test helpers
